Move test/pkg to addons/packages/test/pkg
