+++ addons/packages/calico/test/calico_test.go
@@ -12,8 +12,8 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	"sigs.k8s.io/yaml"
 
-	"github.com/vmware-tanzu/tce/test/pkg/repo"
-	"github.com/vmware-tanzu/tce/test/pkg/ytt"
+	"github.com/vmware-tanzu/tce/addons/packages/test/pkg/repo"
+	"github.com/vmware-tanzu/tce/addons/packages/test/pkg/ytt"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
