+++ addons/packages/external-dns/test/e2e/externaldns_suite_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/vmware-tanzu/tce/test/pkg/utils"
+	"github.com/vmware-tanzu/tce/addons/packages/test/pkg/utils"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
