+++ addons/packages/test/pkg/go.mod
@@ -1,0 +1,3 @@
+module github.com/vmware-tanzu/tce/addons/pacakges/test/pkg
+
+go 1.16
