--- test/pkg/go.mod
@@ -1,3 +1,0 @@
-module github.com/vmware-tanzu/tce/test/pkg
-
-go 1.16
